experiments/validation/requirements.txt
good-egg
scikit-learn
pandas
pyarrow
seaborn
matplotlib
scipy
statsmodels
google-generativeai
pydantic
click
pyyaml
